♻️ Move settings page component & fix up styling for last.fm setting
